requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
assemblyai==0.30.0
certifi==2024.7.4
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.2
websockets==12.0
